+++ UlalaAPI/Controllers/FavorisController.cs
@@ -78,9 +78,14 @@ namespace UlalaAPI.Controllers
         /// Delete API/Favori/{id}
         /// </summary>
         /// <param name="id">id du Favori à supprimer</param>
-        public void Delete(int id)
+        public IHttpActionResult Delete(int id)
         {
-            repo.Delete(id);
+            if (repo.GetOne(id) == null) return NotFound();
+            else
+            {
+                repo.Delete(id);
+                return Ok();
+            }
         }
         #endregion
 
@@ -92,7 +97,8 @@ namespace UlalaAPI.Controllers
         /// <param name="Id">Id du Favori à modifier</param>
         public IHttpActionResult Put(int Id, FavoriModel Favori)
         {
-            if (Favori == null || Favori.Utilisateur.Id == 0 || Favori.Enregistrement.Id == 0 || Id == 0 || repo.GetOne(Id)?.MapTo<JouetModel>() == null) return BadRequest();
+            if (repo.GetOne(Id) == null) return NotFound();
+            else if (Favori == null || Favori.Utilisateur.Id == 0 || Favori.Enregistrement.Id == 0) return BadRequest();
             else
             {
                 repo.Update(Id, Favori.ToEntity());
